Feature/book service

+++ src/main/java/ru/dynamica/library/service/BookService.java
@@ -9,5 +9,5 @@ public interface BookService {
 
     BookDto addBook(BookDto newBook);
 
-    BookDto updateBook(int id, String attribute, String newValue);
+    BookDto updateBook(Long id, String attribute, String newValue);
 }
